Prevent NPE

+++ co.migueljimenez.devops.mart.infrastructure/src/main/java/co/migueljimenez/devops/mart/infrastructure/operations/RemoveResource.xtend
@@ -80,7 +80,8 @@ class RemoveResource extends AbstractOperation {
 				// We have to use the id to find the group, as the name is not available anymore
 				// i.e., name is actually the id
 				eObject = infrastructure.model.securityGroups.findFirst[s|name.equals(s.id)]
-				commands.add('''terraform state rm openstack_compute_secgroup_v2.«(eObject as SecurityGroup).name»''')
+				if (eObject !== null)
+					commands.add('''terraform state rm openstack_compute_secgroup_v2.«(eObject as SecurityGroup).name»''')
 			}
 			default:
 				throw new UnsupportedOperationException('''Resource not supported «type»''')
